lab3/cloud_functions/api_handler/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
flask==3.*